+++ PhotoshopTest/src/TestGrayscale/TestGrayscale.cpp
@@ -32,7 +32,7 @@ TEST_CASE("Create Grayscale File 8-bit psd")
 		.height = height,
 		.colormode = colormode,
 	};
-	std::unordered_map<int16_t, std::vector<type>> img_data = {
+	std::unordered_map<int, std::vector<type>> img_data = {
 		{0, std::vector<type>(width * height, value)},
 		{-1, std::vector<type>(width * height, value)},
 	};
@@ -68,7 +68,7 @@ TEST_CASE("Create Grayscale File 8-bit psb")
 		.height = height,
 		.colormode = colormode,
 	};
-	std::unordered_map<int16_t, std::vector<type>> img_data = {
+	std::unordered_map<int, std::vector<type>> img_data = {
 		{0, std::vector<type>(width * height, value)},
 		{-1, std::vector<type>(width * height, value)},
 	};
@@ -104,7 +104,7 @@ TEST_CASE("Create Grayscale File 16-bit psd")
 		.height = height,
 		.colormode = colormode,
 	};
-	std::unordered_map<int16_t, std::vector<type>> img_data = {
+	std::unordered_map<int, std::vector<type>> img_data = {
 		{0, std::vector<type>(width * height, value)},
 		{-1, std::vector<type>(width * height, value)},
 	};
@@ -140,7 +140,7 @@ TEST_CASE("Create Grayscale File 16-bit psb")
 		.height = height,
 		.colormode = colormode,
 	};
-	std::unordered_map<int16_t, std::vector<type>> img_data = {
+	std::unordered_map<int, std::vector<type>> img_data = {
 		{0, std::vector<type>(width * height, value)},
 		{-1, std::vector<type>(width * height, value)},
 	};
@@ -176,7 +176,7 @@ TEST_CASE("Create Grayscale File 32-bit psd")
 		.height = height,
 		.colormode = colormode,
 	};
-	std::unordered_map<int16_t, std::vector<type>> img_data = {
+	std::unordered_map<int, std::vector<type>> img_data = {
 		{0, std::vector<type>(width * height, value)},
 		{-1, std::vector<type>(width * height, value)},
 	};
@@ -212,7 +212,7 @@ TEST_CASE("Create Grayscale File 32-bit psb")
 		.height = height,
 		.colormode = colormode,
 	};
-	std::unordered_map<int16_t, std::vector<type>> img_data = {
+	std::unordered_map<int, std::vector<type>> img_data = {
 		{0, std::vector<type>(width * height, value)},
 		{-1, std::vector<type>(width * height, value)},
 	};
